The version of ts-jest has been upgraded.
Due to the TypeScript version upgrade, the following warning was displayed in `$ npm test`.

```
ts-jest[versions] (WARN) Version 4.0.5 of typescript installed has not been tested with ts-jest. If you're experiencing issues, consider using a supported version (>=3.8.0 <4.0.0-0). Please do not report issues in ts-jest if you are using unsupported versions.
```

This commit fix this.